// lib/api/generated/models/changeDefaultWorkspaceRequest.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Heymoa API
 * Heymoa 서버 REST API
 * OpenAPI spec version: 1.0.0
 */

export interface ChangeDefaultWorkspaceRequest {
  /**
   * 워크스페이스의 13자리 TSID
   * @minLength 13
   * @maxLength 13
   * @pattern ^[0-9A-HJKMNP-TV-Z]{13}$
   */
  workspaceId: string;
}